Change config file location
Move from current directory to home directory.
